+++ packages/sharedlib-test-py/meta.yaml
@@ -8,6 +8,11 @@ requirements:
 source:
   path: src
 build:
+  cross-script: |
+    pwd
+    echo PYTHONPATH: $PYTHONPATH
+    echo _PYTHON_HOST_PLATFORM: $_PYTHON_HOST_PLATFORM
+    echo PYTHONPATH: $PYTHONPATH
   cflags: |
     -I$(PYODIDE_ROOT)/packages/sharedlib-test/src/include
 test:
